vector-db-similarity-search/requirements.txt
Chroma
langchain-ollama
langchain-community
langchain
pypdf
sentence-transformers